requirements.txt
Django==2.0.2
django-bootstrap4==0.0.6
django-localflavor==2.0
django-semanticui-forms==1.6.5
pkg-resources==0.0.0
pytz==2018.3
